refactor: move manager/providers
