+++ app/src/main/java/com/xinto/opencord/di/ManagerModule.kt
@@ -2,6 +2,7 @@ package com.xinto.opencord.di
 
 import android.content.Context
 import com.xinto.opencord.domain.manager.*
+import com.xinto.opencord.manager.*
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
