+++ app/src/main/java/com/xinto/opencord/ui/viewmodel/ChannelPinsViewModel.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.mutableStateMapOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.viewModelScope
-import com.xinto.opencord.domain.manager.PersistentDataManager
+import com.xinto.opencord.manager.PersistentDataManager
 import com.xinto.opencord.domain.model.DomainMessage
 import com.xinto.opencord.store.MessageStore
 import com.xinto.opencord.ui.viewmodel.base.BasePersistenceViewModel
