+++ app/src/main/java/com/xinto/opencord/ui/viewmodel/ChannelsViewModel.kt
@@ -2,7 +2,7 @@ package com.xinto.opencord.ui.viewmodel
 
 import androidx.compose.runtime.*
 import androidx.lifecycle.viewModelScope
-import com.xinto.opencord.domain.manager.PersistentDataManager
+import com.xinto.opencord.manager.PersistentDataManager
 import com.xinto.opencord.domain.model.DomainChannel
 import com.xinto.opencord.store.ChannelStore
 import com.xinto.opencord.store.GuildStore
