+++ app/src/main/java/com/xinto/opencord/ui/viewmodel/ChatViewModel.kt
@@ -3,7 +3,7 @@ package com.xinto.opencord.ui.viewmodel
 import androidx.compose.runtime.*
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.xinto.opencord.domain.manager.PersistentDataManager
+import com.xinto.opencord.manager.PersistentDataManager
 import com.xinto.opencord.domain.model.DomainMessage
 import com.xinto.opencord.rest.body.MessageBody
 import com.xinto.opencord.rest.service.DiscordApiService
