+++ app/src/main/java/com/xinto/opencord/ui/viewmodel/MembersViewModel.kt
@@ -1,7 +1,7 @@
 package com.xinto.opencord.ui.viewmodel
 
 import androidx.compose.runtime.mutableStateListOf
-import com.xinto.opencord.domain.manager.PersistentDataManager
+import com.xinto.opencord.manager.PersistentDataManager
 import com.xinto.opencord.domain.model.DomainGuildMember
 import com.xinto.opencord.ui.viewmodel.base.BasePersistenceViewModel
