+++ app/src/main/java/com/xinto/opencord/ui/viewmodel/base/BasePersistenceViewModel.kt
@@ -1,7 +1,7 @@
 package com.xinto.opencord.ui.viewmodel.base
 
 import androidx.lifecycle.ViewModel
-import com.xinto.opencord.domain.manager.PersistentDataManager
+import com.xinto.opencord.manager.PersistentDataManager
 
 abstract class BasePersistenceViewModel(
     private val persistentDataManager: PersistentDataManager
